+++ tft_mas/CfgVehicleClasses.hpp
@@ -1,0 +1,12 @@
+class mas_ind_stat {
+    displayName = "Static (MASSI)";
+};
+class mas_ind_veh {
+    displayName = "Cars (MASSI)";
+};
+class mas_ind_arm {
+    displayName = "Armored (MASSI)";
+};
+class mas_ind_air {
+    displayName = "Air (MASSI)";
+};
